requirements.txt
PyYAML>=6.0.1
rich>=13.7.1
watchdog>=4.0.0
cryptography>=43.0.0
xxhash>=3.4.1
